+++ .gitignore
@@ -15,3 +15,5 @@
 # vendor/
 
 *.env
+
+*/static/*
